Expose bindings to integrate with other WGPU capabilities
For my project I need to have some interaction of Burn and other WGPU based frameworks (eg. egui, maybe bevy). This is a hacky draft to see if this might be viable.

The setup requires two parts:

Firstly, exposing a way to run custom commands on the WgpuServer. This is done with a mechanism that gets a callback with exclusive access to the server, and a few handes that are resoved to their underlying resource. The server can then expose some custom commands, or in this case I'm just exposing the CommandEncoder directly. Since this requires exclusive access, I think? this is safe. WASM might need some massaging depending on how Sync/Send WGPU types are there, not sure yet... I also ran into a problem trying this approach for the MSPC ComputeChannel, it requires a static lifetime on the closure for reasons I don't quite understand. Maybe some different approach is needed here!

Secondly, creating a new "Existing" device as WgpuDevice. I don't think it's possible to send data from one wgpu instance/device to another, so I need to be able to have the WGPU backend use the exact existing wgpu devices. I think ideally this would maybe look something like

enum WgpuDevice { Discrete, Integrated, ..., Existing(Arc<Adapter>, Arc<Device>, Arc<Queue>) }

But that doesn't quite work with WgpuDevice, as it needs be Hash etc. Instead I'm now requiring to call wgpu::init_existing_device, which registers the WgpuDevice. It is still possible to register multiple devices like that.